+++ newlib/libc/string/memrchr.c
@@ -33,9 +33,10 @@ QUICKREF
 #include <_ansi.h>
 #include <string.h>
 #include <limits.h>
+#include <stdint.h>
 
 /* Nonzero if X is not aligned on a "long" boundary.  */
-#define UNALIGNED(X) ((long)(X + 1) & (sizeof (long) - 1))
+#define UNALIGNED(X) ((uintptr_t)(X + 1) & (sizeof (long) - 1))
 
 /* How many bytes are loaded each iteration of the word copy loop.  */
 #define LBLOCKSIZE (sizeof (long))
