+++ newlib/libc/string/memset.c
@@ -44,9 +44,10 @@ QUICKREF
 
 #include <string.h>
 #include "local.h"
+#include <stdint.h>
 
 #define LBLOCKSIZE (sizeof(long))
-#define UNALIGNED(X)   ((long)X & (LBLOCKSIZE - 1))
+#define UNALIGNED(X)   ((uintptr_t)X & (LBLOCKSIZE - 1))
 #define TOO_SMALL(LEN) ((LEN) < LBLOCKSIZE)
 
 #undef memset
