+++ newlib/libc/string/strlen.c
@@ -45,9 +45,10 @@ QUICKREF
 #include <_ansi.h>
 #include <string.h>
 #include <limits.h>
+#include <stdint.h>
 
 #define LBLOCKSIZE   (sizeof (long))
-#define UNALIGNED(X) ((long)X & (LBLOCKSIZE - 1))
+#define UNALIGNED(X) ((uintptr_t)X & (LBLOCKSIZE - 1))
 
 #if LONG_MAX == 2147483647L
 #define DETECTNULL(X) (((X) - 0x01010101) & ~(X) & 0x80808080)
